refactor: 쓰임 없는 클래스 삭제 및 dto refactor
경로 수정: domain.auth.kakao.dto.response -> client.kakao.dto
